--- weathereye/playbooks/project/wx_django/templates/wx_web_app/shutdown.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="x-ua-compatible" content="ie=edge">
-
-  <link rel="icon" type="image/png" href="{% static 'images/circle_logo_small.png' %}">
-  <link rel="stylesheet" href="{% static 'css/style.css' %}">
-  
-  <title>Shutdown</title>
-</head>
-<body>
-  <div class="logo"></div>
-
-  <section>
-    <h1 class="shutdown-h1">SURFACE configuration site is offline</h1>
-
-    <p class="shutdown-p">You can now close this page!</p>
-  </section>
-
-  <script>
-    // Wait for a few seconds before shutting down
-    setTimeout(function() {
-        window.location.href = '/shutdown-server/';
-    }, 3000); // 3000 milliseconds = 3 seconds
-  </script>
-
-</body>
-</html>
-
-
