--- weathereye/playbooks/project/wx_django/templates/wx_web_app/success.html
@@ -1,116 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="x-ua-compatible" content="ie=edge">
-
-  <link rel="icon" type="image/png" href="{% static 'images/circle_logo_small.png' %}">
-  <link rel="stylesheet" href="{% static 'css/style.css' %}">
-  
-  <title>Configuration Successful</title>
-</head>
-<body>
-  <div class="logo"></div>
-
-  <section>
-    <h1 class="success-h1">Configuration Successful!</h1>
-    <p class="success-p">Your SURFACE environment variables have been saved.</p>
-
-    <h3>Surface Installation in Progress</h3>
-    <div id="progress-container">
-        <div id="progress-bar"><span style="padding-left: 5px; padding-right: 5px;">0%</span></div>
-    </div>
-    <div id="status">Loading...</div>
-
-    <div class="btn-center">
-      <!-- Continue Installation Button -->
-      <a href="{% url 'shutdown' %}" id="shutdownButton" class="button button-red">Exit</a>
-    </div>
-  </section>
-  <script>
-    // Function to click the shutdown button after 5 hours if the site is still running
-    function clickButton() {
-      document.getElementById('shutdownButton').click();
-    }
-
-    const fiveHoursInMilliseconds = 5 * 60 * 60 * 1000;
-
-    setTimeout(clickButton, fiveHoursInMilliseconds);
-
-    // progress bar control
-    async function fetchFileContent() {
-        try {
-            const response = await fetch('/install-progress/');
-            const text = await response.text();
-            return text;
-        } catch (error) {
-            console.error('Error fetching the file:', error);
-            return '';
-        }
-    }
-
-    function updateProgressBar(text) {
-        const length = text.length;
-        const percentage = Math.min(length * 100 / 21, 100);
-        const progressBar = document.getElementById('progress-bar');
-        const status = document.getElementById('status');
-
-        if (text.includes('x')) {
-            progressBar.style.width = '100%';
-            progressBar.style.backgroundColor = 'red';
-            progressBar.textContent = 'Install Failed';
-            status.textContent = 'Install Failed';
-        } else if (text.includes('s')) {
-            progressBar.style.width = '100%';
-            progressBar.style.backgroundColor = '#4caf50';
-            progressBar.textContent = '100%';
-            status.textContent = 'Loading Complete';
-        }else {
-            progressBar.style.backgroundColor = '#4caf50'
-            progressBar.style.width = Math.floor(percentage) + '%';
-            progressBar.textContent = Math.floor(percentage) + '%';
-            if (percentage === 100) {
-                status.textContent = 'Loading Complete';
-            } else {
-                status.textContent = 'Loading...';
-            }
-        }
-    }
-
-    async function checkFile() {
-        const text = await fetchFileContent();
-        updateProgressBar(text);
-    }
-
-    setInterval(checkFile, 10000); // Check every minute (10 seconds)
-
-    checkFile(); // initial check
-
-    // Function to write 'p' to file
-    function writeToFile() {
-      fetch('/installation-progress/')  // Call the endpoint to trigger the file writing
-        .catch(error => console.error('Error:', error));
-    }
-
-    // Schedule writing 'p' to file every 2.647 minutes
-    function scheduleWrite() {
-        let counter = 0;
-        const interval = 2.647 * 60 * 1000; // Convert minutes to milliseconds
-        const writeInterval = setInterval(() => {
-            writeToFile();
-            counter++;
-            if (counter === 17) {
-                clearInterval(writeInterval);
-            }
-        }, interval);
-    }
-
-    scheduleWrite(); // Start the process
-
-  </script>
-</body>
-</html>
-
-
