--- weathereye/playbooks/project/wx_django/wx_django/urls.py
@@ -1,27 +1,0 @@
-"""wx_django URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.contrib import admin
-from django.urls import path
-from wx_web_app import views
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', views.configure_surface, name='configure_surface'),
-    path('shutdown/', views.shutdown, name='shutdown'),
-    path('shutdown-server/', views.shutdown_server, name='shutdown_server'),
-    path('install-progress/', views.get_install_progress, name='get_install_progress'),
-    path('installation-progress/', views.simulate_progress, name='simulate_progress'),
-]
