--- weathereye/playbooks/project/wx_django/wx_web_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class WxWebAppConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'wx_web_app'
