+++ weathereye/wx_config/ansible/surface_app/inventory/hosts
