+++ weathereye/wx_config/ansible/surface_app/project/ansible.cfg
@@ -1,0 +1,57 @@
+[defaults]
+# (boolean) This controls whether an Ansible playbook should prompt for a login password. If using SSH keys for authentication, you probably do not need to change this setting.
+;ask_pass=True
+
+# (path) The password file to use for the become plugin. --become-password-file.
+# If executable, it will be run and the resulting stdout will be used as the password.
+become_password_file= ../env/become_password
+
+# (path) The password file to use for the connection plugin. --connection-password-file.
+connection_password_file= ../env/connection_password
+
+# (integer) Sets the default verbosity, equivalent to the number of ``-v`` passed in the command line.
+verbosity=0
+
+# (boolean) By default Ansible will issue a warning when there are no hosts in the inventory.
+# These warnings can be silenced by adjusting this setting to False.
+localhost_warning=False
+
+# (list) List of enabled callbacks, not all callbacks need enabling, but many of those shipped with Ansible do as we don't want them activated by default.
+callbacks_enabled=community.general.yaml
+
+# (string) Set the main callback used to display Ansible output. You can only have one at a time.
+# You can have many other callbacks, but just one can be in charge of stdout.
+# See :ref:`callback_plugins` for a list of available options.
+stdout_callback=community.general.yaml
+
+# (integer) Set the maximum time (in seconds) that a task can run for.
+# If set to 0 (the default) there is no timeout.
+task_timeout=7200
+
+
+[callback_yml]
+# Output to be used for callback runner_retry.
+runner_retry_msg= Current task: {{ ansible_callback_yml.task.name }}, is still in progress, please wait - Checking again in a few seconds...
+
+# Output color to be used for runner_retry_msg.
+runner_retry_msg_color=cyan
+
+
+[privilege_escalation]
+# (boolean) Display an agnostic become prompt instead of displaying a prompt containing the command line supplied become method
+;agnostic_become_prompt=True
+
+# (boolean) Toggles the use of privilege escalation, allowing you to 'become' another user after login.
+become=True
+
+# (boolean) Toggle to prompt for privilege escalation password.
+;become_ask_pass=True
+
+# (string) The user your login/remote user 'becomes' when using privilege escalation, most systems will use 'root' when no user is specified.
+become_user=root
+
+
+[inventory]
+# (boolean) By default Ansible will issue a warning when no inventory was loaded and notes that it will use an implicit localhost-only inventory.
+# These warnings can be silenced by adjusting this setting to False.
+inventory_unparsed_warning=True
