+++ weathereye/wx_config/config_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ConfigAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'config_app'
