+++ weathereye/wx_config/surface_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class SurfaceAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'surface_app'
