+++ weathereye/wx_playbook/inventory/hosts
@@ -1,0 +1,2 @@
+[local]
+localhost ansible_connection=local
