+++ weathereye/wx_playbook/project/ansible.cfg
