+++ weathereye/wx_playbook/project/wx_configuration.yml
@@ -1,0 +1,55 @@
+---
+- name: Check Supported Systems
+  hosts: local
+  gather_facts: true
+  become: false # granting root privileges
+
+  tasks:
+  # display Users system information
+    - name: System
+      ansible.builtin.debug:
+        msg: The distribution is "{{ ansible_facts['distribution'] }}" and the version is "{{ ansible_facts['distribution_version'] }}"
+      
+  # Users system must be a Mac or Ubuntu system
+    - name: System check
+      ansible.builtin.fail:
+        msg: This host system is currently not supported. See docs.weathereye.org for supported operating systems and distributions.
+      when: ansible_distribution not in ['MacOSX', 'Ubuntu']
+
+
+- name: Configure SURFACE
+  hosts: local
+  become: true # granting root privileges
+
+  tasks:
+  # shutdown old configuration sites that may still be running
+    - name: Shut down stale configuration site
+      ansible.builtin.uri:
+        url: http://localhost:52376/shutdown/
+        method: GET
+        force: true
+        status_code: 200
+      ignore_errors: true
+  
+  # ensures django 4.1 is install
+    - name: Install required packages
+      ansible.builtin.shell: pip install --upgrade django==4.1
+  
+  # starting configuration site
+    - name: Start configuration site
+      ansible.builtin.shell: python3 manage.py runserver localhost:52376
+      args:
+        chdir: "{{ django_webapp_path }}"
+      async: 10800
+      poll: 0
+
+    - name: Wait for the server to start
+      ansible.builtin.pause:
+        seconds: 5
+
+    - name: Reload configuration site
+      ansible.builtin.uri:
+        url: http://localhost:52376/
+        method: GET
+        force: true
+        status_code: 200
